requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1  
cx-Oracle==8.3.0
Django==4.1.6
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
Pillow==9.3.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14